[ENSCORESW-571]. Moved target compilation and sync'd the shell scripts.
Firstly the target of the JAR file and class files have moved under the
/target directory. Target has never been committed into CVS and so is
perfect to be ignored. build and dist are both ignored now by default. The
DiscoveryBasedTestRegistry has been updated accordingly to show
this move.

Secondly the execution lines for the shell scripts are now in sync. This
includes the "web" ones which are actually the production ones.