ignore LRGs for all divisions but vertebrates
In Compara we do not use LRGs for other divisions than vertebrates. In
fact, we have stopped copying then, which is making CheckTopLevelDnaFrag
healthcheck to fail since some LRGs are missing for plants division.
This patch fixes the problem, not comparing LRGs between compara master
and core databases when the division is not vertebrates.